fix: Address comments
