+++ apps/api/index.ts
@@ -96,9 +96,13 @@ app.use(
     },
   })
 );
+
 app.use((err: unknown, req: Request, res: Response, next: NextFunction) => {
+  // add the error to the response so pino-http can log it
+  res.err = err as Error;
+
   if (err instanceof HTTPError) {
-    return res.status(err.code).send({
+    res.status(err.code).send({
       errors: [
         {
           title: err.message,
@@ -107,6 +111,8 @@ app.use((err: unknown, req: Request, res: Response, next: NextFunction) => {
         },
       ],
     });
+
+    return next(err);
   }
 
   return res.status(500).json({
