+++ apps/api/routes/crm/v1/index.ts
@@ -4,6 +4,8 @@ import accounts from './account';
 import contacts from './contact';
 import leads from './lead';
 import opportunities from './opportunity';
+import syncHistory from './sync_history';
+import syncInfo from './sync_info';
 
 export default function init(app: Router): void {
   const v1Router = Router();
@@ -14,6 +16,9 @@ export default function init(app: Router): void {
   accounts(v1Router);
   leads(v1Router);
   opportunities(v1Router);
+
+  syncInfo(v1Router);
+  syncHistory(v1Router);
 
   v1Router.use(openApiErrorHandlerMiddleware);
