+++ apps/api/routes/crm/v1/sync_info.ts
@@ -1,6 +1,6 @@
 import { getDependencyContainer } from '@/dependency_container';
 import { snakecaseKeys } from '@/lib/snakecase';
-import { GetSyncInfosPathParams, GetSyncInfosRequest, GetSyncInfosResponse } from '@supaglue/schemas/customer';
+import { GetSyncInfosPathParams, GetSyncInfosRequest, GetSyncInfosResponse } from '@supaglue/schemas/crm';
 import { Request, Response, Router } from 'express';
 
 const { syncService } = getDependencyContainer();
@@ -14,7 +14,7 @@ export default function init(app: Router): void {
       req: Request<GetSyncInfosPathParams, GetSyncInfosResponse, GetSyncInfosRequest>,
       res: Response<GetSyncInfosResponse>
     ) => {
-      const syncInfoList = await syncService.getSyncInfoList(req.sg.connectionId);
+      const syncInfoList = await syncService.getSyncInfoList(req.customerConnection.id);
       const syncInfoListRes = syncInfoList.map((syncInfo) =>
         snakecaseKeys({
           ...syncInfo,
