+++ apps/api/routes/customer/integration/connection/index.ts
@@ -1,8 +1,6 @@
 import { getDependencyContainer } from '@/dependency_container';
 import { connectionMiddleware } from '@/middleware/connection';
 import { Request, Response, Router } from 'express';
-import syncHistory from './sync_history';
-import syncInfo from './sync_info';
 
 const { connectionService } = getDependencyContainer();
 
@@ -29,9 +27,6 @@ export default function init(app: Router): void {
     const connection = await connectionService.delete(req.sg.connectionId);
     return res.status(200).send(connection);
   });
-
-  syncInfo(connectionRouter);
-  syncHistory(connectionRouter);
 
   app.use('/connections/:connectionId', connectionMiddleware, connectionRouter);
 }
