+++ openapi/common/components/parameters/provider-name.yaml
@@ -1,0 +1,7 @@
+name: provider-name
+in: header
+schema:
+  type: string
+  example: salesforce
+description: The provider name
+required: true
