+++ openapi/crm/components/schemas/sync-history.yaml
