+++ openapi/crm/components/schemas/sync-info.yaml
