+++ openapi/crm/paths/accounts@{account_id}.yaml
@@ -77,6 +77,8 @@ patch:
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
 parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
   - name: account_id
     in: path
     required: true
