+++ openapi/crm/paths/contacts.yaml
@@ -111,3 +111,6 @@ post:
                 $ref: ../components/schemas/objects/contact.yaml
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
+parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
