+++ openapi/crm/paths/leads.yaml
@@ -125,3 +125,6 @@ post:
                 $ref: ../components/schemas/objects/lead.yaml
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
+parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
