+++ openapi/crm/paths/leads@{lead_id}.yaml
@@ -55,6 +55,8 @@ patch:
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
 parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
   - name: lead_id
     in: path
     required: true
