+++ openapi/crm/paths/opportunities.yaml
@@ -99,3 +99,6 @@ post:
                 $ref: ../components/schemas/objects/opportunity.yaml
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
+parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
