+++ openapi/crm/paths/opportunities@{opportunity_id}.yaml
@@ -61,6 +61,8 @@ patch:
               warnings:
                 $ref: ../../common/components/schemas/warnings.yaml
 parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
   - name: opportunity_id
     in: path
     required: true
