+++ openapi/crm/paths/sync-history.yaml
@@ -8,9 +8,6 @@ get:
   parameters:
     - $ref: ../../common/components/parameters/cursor.yaml
     - $ref: ../../common/components/parameters/page_size.yaml
-    - $ref: ../components/parameters/customer_id.yaml
-    - $ref: ../components/parameters/integration_id.yaml
-    - $ref: ../components/parameters/connection_id.yaml
     - name: model
       in: query
       schema:
@@ -45,3 +42,6 @@ get:
                     end_timestamp: 2023-02-22T20:50:17.559537Z
                     status: ERROR
                     error_message: 'Error: Something went wrong'
+parameters:
+  - $ref: ../../common/components/parameters/customer-id.yaml
+  - $ref: ../../common/components/parameters/provider-name.yaml
